Reduced memory overhead; updated existing effects; fixed solid color effect constantly running

+++ src/effects/fade_black.cpp
@@ -2,7 +2,7 @@
 #include <FastLED.h>
 #include "fade_black.h"
 
-void FadeBlack::run(CRGBSet &led_array, uint32_t led_num, uint8_t &serial_mode)
+void FadeBlack::run(CRGBSet &led_array, uint32_t led_num, uint32_t &serial_mode)
 {
     // Set current mode to none
     serial_mode = 0;
